refactor: separate decoder from encoder
