+++ src/decoder/decodable.rs
@@ -1,0 +1,4 @@
+pub trait Decodable {
+    // TODO: implement validate or reader mechanism
+    fn decode(bytes: &[u8]) -> Self;
+}
