+++ src/decoder/mod.rs
@@ -1,0 +1,4 @@
+mod decodable;
+mod impls;
+
+pub use decodable::Decodable;
