+++ src/encoder/impls.rs
@@ -16,19 +16,6 @@ impl Encodable for BitArray {
 
         encoded
     }
-
-    fn decode(bytes: &[u8]) -> Self {
-        let buffer = &bytes[..8];
-        let byte_size = u64::from_be_bytes(buffer.try_into().unwrap()) as usize;
-        let byte_data = &bytes[8..8 + byte_size];
-        let buffer = &bytes[(8 + byte_size)..];
-        let size = u64::from_be_bytes(buffer.try_into().unwrap()) as usize;
-
-        Self {
-            byte_array: byte_data.into(),
-            size,
-        }
-    }
 }
 
 impl Encodable for BloomFilter {
@@ -43,19 +30,6 @@ impl Encodable for BloomFilter {
         encoded.extend_from_slice(&(self.hash_count as u64).to_be_bytes());
 
         encoded
-    }
-
-    fn decode(bytes: &[u8]) -> Self {
-        let split_idx = bytes.len() - 8;
-        let buffer = &bytes[..split_idx];
-        let bit_array = BitArray::decode(buffer);
-        let buffer = &bytes[split_idx..];
-        let hash_count = u64::from_be_bytes(buffer.try_into().unwrap()) as usize;
-
-        Self {
-            bit_array,
-            hash_count,
-        }
     }
 }
 
@@ -94,20 +68,6 @@ mod encodable {
                 ]
             );
         }
-
-        #[test]
-        fn test_decode() {
-            let encoded = vec![
-                0, 0, 0, 0, 0, 0, 0, 2, // Byte size
-                0b00000100, 0b01000000, // Byte data
-                0, 0, 0, 0, 0, 0, 0, 10, // Bin size
-            ];
-            let bit_array = BitArray::decode(&encoded);
-            assert_eq!(bit_array.size, 10);
-            assert!(!bit_array[0]);
-            assert!(bit_array[5]);
-            assert!(bit_array[9]);
-        }
     }
 
     mod bloom_filter {
@@ -143,22 +103,6 @@ mod encodable {
                     0, 0, 0, 0, 0, 0, 0, 7, // Number of hash functions
                 ]
             );
-        }
-
-        #[test]
-        fn test_decode() {
-            let encoded = vec![
-                0, 0, 0, 0, 0, 0, 0, 3, // BitArray: byte size
-                0b10000100, 0b00100001, 0, // BitArray: byte data
-                0, 0, 0, 0, 0, 0, 0, 20, // BitArray: bin size
-                0, 0, 0, 0, 0, 0, 0, 7, // Number of hash functions
-            ];
-            let bloom_filter = BloomFilter::decode(&encoded);
-            assert_eq!(bloom_filter.bit_array.byte_array.len(), 3);
-            assert_eq!(bloom_filter.bit_array.size, 20);
-            assert_eq!(bloom_filter.hash_count, 7);
-            assert!(bloom_filter.lookup("test"));
-            assert!(!bloom_filter.lookup("test1"));
         }
     }
 }
