+++ src/lib.rs
@@ -1,5 +1,6 @@
 mod bit_array;
 mod bloom_filter;
+mod decoder;
 mod encoder;
 mod hash;
